size_t warning in gamepacketnew fixed

+++ src/common/Network/GamePacketNew.h
@@ -108,7 +108,7 @@ namespace Sapphire::Network::Packets
     uint16_t m_segmentType;
 
   public:
-    virtual uint32_t getContentSize()
+    virtual size_t getContentSize()
     {
       return 0;
     };
@@ -195,7 +195,7 @@ namespace Sapphire::Network::Packets
     FFXIVIpcPacket< T, T1 >( const FFXIVARR_PACKET_RAW& rawPacket )
     {
       auto ipcHdrSize = sizeof( FFXIVARR_IPC_HEADER );
-      auto copySize = std::min< uint32_t >( sizeof( T ), rawPacket.segHdr.size - ipcHdrSize );
+      auto copySize = std::min< size_t >( sizeof( T ), rawPacket.segHdr.size - ipcHdrSize );
 
       memcpy( &m_segHdr, &rawPacket.segHdr, sizeof( FFXIVARR_PACKET_SEGMENT_HEADER ) );
       memcpy( &m_data, &rawPacket.data[ 0 ] + ipcHdrSize, copySize );
@@ -204,7 +204,7 @@ namespace Sapphire::Network::Packets
       m_ipcHdr.type = static_cast< ServerZoneIpcType >( m_data._ServerIpcType );
     }
 
-    uint32_t getContentSize() override
+    size_t getContentSize() override
     {
       return sizeof( FFXIVARR_IPC_HEADER ) + sizeof( T );
     }
@@ -291,7 +291,7 @@ namespace Sapphire::Network::Packets
       memcpy( &m_segHdr, data, segmentHdrSize );
     }
 
-    uint32_t getContentSize() override
+    size_t getContentSize() override
     {
       return m_data.size();
     }
